feat(scan): improve scanning performance by conditionally skipping type determination
Determinig the type of a file involves making a call to ffprobe.
For larger libraries, this can incur a significant performance penalty on every scan.

To get around this, this change moves the filetype determination in scan(), so
that the type is only scanned for new or changed files. If the file has remained the same,
we can just reuse the old filetype.